[feat] #28 K,M,B 숫자 단위 적용

+++ app/src/main/java/com/android/go/sopt/winey/domain/entity/WineyFeedModel.kt
@@ -2,9 +2,9 @@ package com.android.go.sopt.winey.domain.entity
 
 import androidx.annotation.DrawableRes
 
-class WineyFeedModel (
+class WineyFeedModel(
     val feedMoney: Int,
-    var likes: Int,
+    var likes: Long,
     @DrawableRes val feedImage: Int,
     var feedTitle: String,
     var nickName: String,
